Move some code to module_utils.
